test(landmark,map,outpost): migrate/rename real-I/O unit tests onto the seam
- map/exporter: the Exporter takes its fs as a constructor arg, so pass
  createMockFs and read the materialised pathway tree back through it (fixed
  /map-export root). Asserts pure HTML/RDF output.
- map/activity/substrate-stage: phase-ordering tests already used a mock
  runtime; the "explicit target" test only threads a path through, so it now
  uses a fixed string. The two real-fs cases (copy-activity ENOENT via real
  fs.cp, bootstrap-shape parity comparing two materialised trees) move to
  substrate-stage.integration.test.js.
- landmark/dispatcher and lib/commands-verb: spawn the real bin (`node`/`bunx`)
  to assert process-level exit codes / argv branch — rename to
  *.integration.test.js and drop them from check-subprocess-in-tests.deny.json
  (now whole-file exempt as integration smoke tests; deny-list is empty).
- map/pipeline: real GraphIndex over LocalStorage + mkdtemp — rename to
  *.integration.test.js.
- outpost/sync-helpers-copy: copyThreadAttachments is a skill .mjs using real
  node:fs with no injection seam; its path-traversal containment check is a
  real-fs property — rename to *.integration.test.js.

All touched packages already declare @forwardimpact/libmock.

Refs spec 0640 § B (Part 02 sweep).
