+++ products/map/test/pipeline.integration.test.js
