+++ scripts/check-subprocess-in-tests.deny.json
@@ -1,4 +1,1 @@
-[
-  "products/landmark/test/dispatcher.test.js",
-  "products/landmark/test/lib/commands-verb.test.js"
-]
+[]
